+++ tests/test_cli/test_dcos_aws/test_cli.py
@@ -1,0 +1,163 @@
+"""
+Tests for the AWS CLI.
+
+This mostly provides error case coverage.
+We rely mostly on manual testing.
+This is because automated tests for this would be very slow.
+
+For developing help texts, it is useful to add a breakpoint on failure and then
+to capture what the help text actually is with:
+
+  .. code: python
+
+       import pyperclip; pyperclip.copy(result.output)
+"""
+
+from textwrap import dedent
+from typing import List
+
+import pytest
+from click.testing import CliRunner
+
+from cli import dcos_aws
+
+
+class TestDcosAWS:
+    """
+    Tests for the top level `dcos-aws` command.
+    """
+
+    def test_version(self) -> None:
+        """
+        The CLI version is shown with ``dcos-aws --version``.
+        """
+        runner = CliRunner()
+        result = runner.invoke(
+            dcos_aws,
+            ['--version'],
+            catch_exceptions=False,
+        )
+
+        assert result.exit_code == 0
+        expected = 'dcos-aws, version'
+        assert expected in result.output
+
+    @pytest.mark.parametrize('arguments', [
+        [],
+        ['--help'],
+    ])
+    def test_help(self, arguments: List[str]) -> None:
+        """
+        Help test is shown with `dcos-aws` and `dcos-aws --help`.
+        """
+        runner = CliRunner()
+        result = runner.invoke(dcos_aws, arguments, catch_exceptions=False)
+        assert result.exit_code == 0
+        # yapf breaks multi-line noqa, see
+        # https://github.com/google/yapf/issues/524.
+        # yapf: disable
+        expected_help = dedent(
+            """\
+            Usage: dcos-aws [OPTIONS] COMMAND [ARGS]...
+
+              Manage DC/OS clusters on AWS.
+
+            Options:
+              --version  Show the version and exit.
+              --help     Show this message and exit.
+
+            Commands:
+              create  Create a DC/OS cluster.
+            """,# noqa: E501,E261
+        )
+        # yapf: enable
+        assert result.output == expected_help
+
+
+class TestCreate:
+    """
+    Tests for the `create` subcommand.
+    """
+
+    def test_help(self) -> None:
+        """
+        Help text is shown with `dcos-aws create --help`.
+        """
+        runner = CliRunner()
+        result = runner.invoke(
+            dcos_aws,
+            ['create', '--help'],
+            catch_exceptions=False,
+        )
+        assert result.exit_code == 0
+        # yapf breaks multi-line noqa, see
+        # https://github.com/google/yapf/issues/524.
+        # yapf: disable
+        expected_help = dedent(
+            """\
+            Usage: dcos-aws create [OPTIONS]
+
+              Create a DC/OS cluster.
+
+                  DC/OS Enterprise
+
+                              DC/OS Enterprise clusters require different configuration variables to DC/OS OSS.
+                              For example, enterprise clusters require the following configuration parameters:
+
+                      ``superuser_username``, ``superuser_password_hash``,
+                      ``fault_domain_enabled``, ``license_key_contents``
+
+                              These can all be set in ``--extra-config``.
+                              However, some defaults are provided for all but the license key.
+
+                              The default superuser username is ``admin``.
+                              The default superuser password is ``admin``.
+                              The default ``fault_domain_enabled`` is ``false``.
+
+                              ``license_key_contents`` must be set for DC/OS Enterprise 1.11 and above.
+                              This is set to one of the following, in order:
+
+                              * The ``license_key_contents`` set in ``--extra-config``.
+                              * The contents of the path given with ``--license-key``.
+                              * The contents of the path set in the ``DCOS_LICENSE_KEY_PATH`` environment variable.
+
+                              If none of these are set, ``license_key_contents`` is not given.
+
+            Options:
+              --variant [oss|enterprise]      Choose the DC/OS variant. If the variant does
+                                              not match the variant of the given artifact
+                                              URL, an error will occur.
+              --masters INTEGER               The number of master nodes.  [default: 1]
+              --agents INTEGER                The number of agent nodes.  [default: 1]
+              --extra-config PATH             The path to a file including DC/OS
+                                              configuration YAML. The contents of this file
+                                              will be added to add to a default
+                                              configuration.
+              --public-agents INTEGER         The number of public agent nodes.  [default:
+                                              1]
+              --workspace-dir PATH            Creating a cluster can use approximately 2 GB
+                                              of temporary storage. Set this option to use a
+                                              custom "workspace" for this temporary storage.
+                                              See https://docs.python.org/3/library/tempfile
+                                              .html#tempfile.gettempdir for details on the
+                                              temporary directory location if this option is
+                                              not set.
+              --license-key PATH              This is ignored if using open source DC/OS. If
+                                              using DC/OS Enterprise, this defaults to the
+                                              value of the `DCOS_LICENSE_KEY_PATH`
+                                              environment variable.
+              --security-mode [disabled|permissive|strict]
+                                              The security mode to use for a DC/OS
+                                              Enterprise cluster. This overrides any
+                                              security mode set in ``--extra-config``.
+              --copy-to-master TEXT           Files to copy to master nodes before
+                                              installing DC/OS. This option can be given
+                                              multiple times. Each option should be in the
+                                              format /absolute/local/path:/remote/path.
+              -v, --verbose                   Use verbose output. Use this option multiple
+                                              times for more verbose output.
+              --help                          Show this message and exit.
+            """,# noqa: E501,E261
+        )
+        # yapf: enable
+        assert result.output == expected_help